Add working directory and refactor BxtConfig methods
In BxtConfig module, a working directory is now created if it does not exist, and is also saved in the configuration. Method names have been refactored for consistency, and the user input method for configuration options has been converted into a static method. These changes enhance the organization of user configurations and improve code readability.