Add 'access_token' to session cookies in scripts
In an attempt to deduce the name for the token stored in the cookieJar. In all four scripts, the 'token' was added to the session cookies with both 'token' and 'access_token' names. Additionally, the cookie jar log messages were updated to show this change. This has been done to allow for authentication using either of these cookie names.